+++ publish/CHANGELOG.md
@@ -1,2 +1,11 @@
+- v1.1
+  - Adds support for multiple ids in the prefab field.
+  - Adds fields events and eventDistance.
+  - Adds fields objects and objectDistance.
+  - Adds fields locations and locationDistance.
+  - Adds fields day and night.
+  - Changes weight calculation to allow nothing if total weight is less than 1.
+  - Fixes command not working without data or swap.
+
 - v1.0
   - Initial release.